requirements.txt
flask
flask-caching
requests
python-dotenv

